requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
# python-dotenv is optional now since we're using local Ollama
